led-api-server/pyproject.toml
[tool.poetry]
name = "led-api-server"
version = "0.1.0"
description = ""
authors = ["HarmvZ <harm.van.zeeland@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"
uWSGI = "^2.0.19"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
